dhkem: rename `arithmetic` feature to `ecdh`
This maps to the `ecdh` feature name of the respective crates
