+++ dhkem/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "dhkem"
 description = """
-Key Encapsulation Mechanism (KEM) adapters for Elliptic Curve Diffie Hellman (ECDH) protocols
+Pure Rust implementation of Key Encapsulation Mechanism (KEM) adapters for Elliptic Curve
+Diffie Hellman (ECDH) protocols
 """
 version = "0.1.0"
 edition = "2021"
@@ -15,6 +16,8 @@ readme = "README.md"
 [dependencies]
 kem = "0.3.0-pre.0"
 rand_core = "0.6.4"
+
+# optional dependencies
 x25519 = { version = "2.0.1", package = "x25519-dalek", optional = true, default-features = false }
 elliptic-curve = { version = "0.13.8", optional = true, default-features = false }
 bign256 = { version = "0.13.1", optional = true, default-features = false, features = ["arithmetic"] }
@@ -29,16 +32,16 @@ zeroize = { version = "1.8.1", optional = true, default-features = false }
 
 [features]
 default = ["zeroize"]
-arithmetic = ["dep:elliptic-curve", "elliptic-curve/ecdh"]
+ecdh = ["dep:elliptic-curve", "elliptic-curve/ecdh"]
 x25519 = ["dep:x25519", "x25519/reusable_secrets"]
-bign256 = ["dep:bign256", "arithmetic"]
-k256 = ["dep:k256", "arithmetic"]
-p192 = ["dep:p192", "arithmetic"]
-p224 = ["dep:p224", "arithmetic"]
-p256 = ["dep:p256", "arithmetic"]
-p384 = ["dep:p384", "arithmetic"]
-p521 = ["dep:p521", "arithmetic"]
-sm2 = ["dep:sm2", "arithmetic"]
+bign256 = ["dep:bign256", "ecdh"]
+k256 = ["dep:k256", "ecdh"]
+p192 = ["dep:p192", "ecdh"]
+p224 = ["dep:p224", "ecdh"]
+p256 = ["dep:p256", "ecdh"]
+p384 = ["dep:p384", "ecdh"]
+p521 = ["dep:p521", "ecdh"]
+sm2 = ["dep:sm2", "ecdh"]
 zeroize = ["dep:zeroize"]
 
 [dev-dependencies]
